+++ packages/apps/src/types/plugin/decorators/index.ts
@@ -1,0 +1,2 @@
+export { Plugin, PluginOptions } from './plugin';
+export { Inject, InjectOptions, Logger, Storage } from './inject';
