+++ packages/apps/src/types/plugin/index.ts
@@ -8,6 +8,4 @@ export * from './plugin-init-event';
 export * from './plugin-start-event';
 export * from './plugin-stop-event';
 export * from './sender';
-
-export { Plugin, PluginOptions } from './plugin-decorator';
-export { Inject, InjectOptions, Logger, Storage } from './inject-decorator';
+export * from './decorators';
